requirements.txt
APScheduler==3.10.1
asgiref==3.6.0
attrs==22.2.0
Automat==22.10.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
cryptography==39.0.2
cssselect==1.2.0
Django==4.1.7
django-filter==23.1
djangorestframework==3.14.0
fake-useragent==1.1.3
Faker==18.3.0
filelock==3.10.0
Flask==2.2.3
Flask-Compress==1.13
Flask-SQLAlchemy==3.0.3
free-proxy==1.1.1
future==0.18.3
geographiclib==2.0
geopy==2.3.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
logparser==0.8.2
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
packaging==23.0
parsel==1.7.0
pexpect==4.8.0
pproxy==2.7.8
probableparsing==0.0.1
Protego==0.2.1
protobuf==3.20.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.0.0
python-crfsuite==0.9.9
python-dateutil==2.8.2
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
Scrapy==2.8.0
scrapy-inline-requests==0.3.1
scrapy-rotating-proxies==0.6.2
scrapyd==1.4.1
scrapyd-client==1.2.3
scrapydweb==1.4.0
service-identity==21.1.0
six==1.16.0
SQLAlchemy==2.0.7
sqlparse==0.4.3
tldextract==3.4.0
Twisted==22.10.0
typing==3.7.4.3
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uberegg==0.1.1
urllib3==1.26.15
usaddress==0.5.10
w3lib==2.1.1
Werkzeug==2.2.3
zope.interface==6.0
